feat: 유틸함수 분리 및 이전 필요없는 파일 삭제
