+++ src/components/recommend/Recommend.tsx
@@ -13,11 +13,19 @@ const Recommend: React.FC<RecommendProps> = ({ data, MBTIResult }) => {
   return (
     <div className="flex justify-center">
       {data.length > 0 ? (
-        <ul className="grid grid-cols-1 md:grid-cols-3 gap-4 text-[18px] text-grey-600 font-normal">
+        <ul className="grid grid-cols-3 gap-11 text-[18px] text-grey-600 font-normal">
           {data.map((item, index) => (
-            <li key={index} className="flex flex-col items-center space-y-2">
-              <h2 className="text-center">{item.title}</h2>
-              <button onClick={() => router.push(`/detail/${item.contentid}`)}>
+            <li key={index} className="flex flex-col items-center space-y-2 items-stretch">
+              <h2
+                className="text-center w-full truncate"
+                style={{ maxWidth: "120px", whiteSpace: "nowrap", overflow: "hidden", textOverflow: "ellipsis" }}
+              >
+                {item.title}
+              </h2>
+              <button
+                onClick={() => router.push(`/detail/${item.contentid}`)}
+                className="flex-grow flex items-center justify-center"
+              >
                 {item.firstimage ? (
                   <Image src={item.firstimage} alt={item.title} width={100} height={100} className="object-contain" />
                 ) : (
